+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\User;
 use App\Program;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Validator;
 
 
@@ -13,10 +14,6 @@ class UserController extends Controller
     public function __construct()
     {
         $this->middleware('auth:admin');
-        $this->middleware('permission:user-edit', ['only' => ['edit','update']]);
-        $this->middleware('permission:user-list', ['only' => ['show']]);
-        $this->middleware('permission:program-edit', ['only' => ['assignProgram']]);
-        $this->middleware('permission:program-delete', ['only' => ['removeProgram']]);
     }
 
     /**
@@ -58,20 +55,27 @@ class UserController extends Controller
      */
     public function show($id)
     {
-        $id = filter_var($id, FILTER_SANITIZE_NUMBER_INT);
-        $user = User::findOrFail($id);
-        return view('admin.users.show', compact('user'));
+        if (Auth::user()->hasPermissionTo('user-show')) {
+            $id = filter_var($id, FILTER_SANITIZE_NUMBER_INT);
+            $user = User::findOrFail($id);
+            return view('admin.users.show', compact('user'));
+        } else {
+            return redirect()->back()->with('error', 'У Вас нет прав для выполнения этой операции');
+        }
     }
 
 
     public function edit(Request $request)
     {
-        //
-        $id = filter_var($request->id, FILTER_SANITIZE_NUMBER_INT);
-        $user = User::findOrFail($id);
-        $userPrograms = $user->programs()->get();
-        $programs = Program::all();
-        return view('admin.users.edit', compact('user','programs','userPrograms'));
+        if (Auth::user()->hasPermissionTo('user-edit')) {
+            $id = filter_var($request->id, FILTER_SANITIZE_NUMBER_INT);
+            $user = User::findOrFail($id);
+            $userPrograms = $user->programs()->get();
+            $programs = Program::all();
+            return view('admin.users.edit', compact('user','programs','userPrograms'));
+        } else {
+            return redirect()->back()->with('error', 'У Вас нет прав для выполнения этой операции');
+        }
     }
 
     /**
@@ -83,7 +87,6 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        dd($request->all());  
         $validator = Validator::make($request->all(), [
             'name' => 'string|max:255',
             'phone' => 'string|min:9',
@@ -118,33 +121,44 @@ class UserController extends Controller
      */
     public function destroy(Request $request)
     {
-        $userId = filter_var($request->id, FILTER_SANITIZE_NUMBER_INT);
-        $user = User::findOrFail($userId);
-        if($user->image) unlink(storage_path('app/public/'.$user->image));
-        $user->delete();
-        return redirect()->back()->with('success','Пользователь ' . $user->name . ' был успешно удален!');
+        if (Auth::user()->hasPermissionTo('user-delete')) {
+            $userId = filter_var($request->id, FILTER_SANITIZE_NUMBER_INT);
+            $user = User::findOrFail($userId);
+            if($user->image) unlink(storage_path('app/public/'.$user->image));
+            $user->delete();
+            return redirect()->back()->with('success','Пользователь ' . $user->name . ' был успешно удален!');
+        } else {
+            return redirect()->back()->with('error', 'У Вас нет прав для выполнения этой операции');
+        }
     }
 
     public function assignProgram(Request $request)
     {
-        $userId = filter_var($request->user, FILTER_SANITIZE_NUMBER_INT);
-        $program = filter_var($request->program, FILTER_SANITIZE_SPECIAL_CHARS);
-        $programs = User::findOrFail($userId)->programs()->get();
-        $programExist = $programs->where('id',$program);
-        if(!count($programExist) > 0){
-            User::findOrFail($userId)->programs()->attach($program);
-            return redirect()->back()->with('success', 'Программа была успешно добавлена!');
+        if (Auth::user()->hasPermissionTo('program-assign')) {
+            $userId = filter_var($request->user, FILTER_SANITIZE_NUMBER_INT);
+            $program = filter_var($request->program, FILTER_SANITIZE_SPECIAL_CHARS);
+            $programs = User::findOrFail($userId)->programs()->get();
+            $programExist = $programs->where('id',$program);
+            if(!count($programExist) > 0){
+                User::findOrFail($userId)->programs()->attach($program);
+                return redirect()->back()->with('success', 'Программа была успешно добавлена!');
+            } else {
+                return redirect()->back()->with('error', 'Пользователью уже была добавлена программа!');
+            }
         } else {
-            return redirect()->back()->with('error', 'Пользователью уже была добавлена программа!');
+            return redirect()->back()->with('error', 'У Вас нет прав для выполнения этой операции');
         }
-
     }
 
     public function removeProgram(Request $request)
     {
-        $userId = filter_var($request->user, FILTER_SANITIZE_NUMBER_INT);
-        $program = filter_var($request->program, FILTER_SANITIZE_SPECIAL_CHARS);
-        $user = User::findOrFail($userId)->programs()->detach($program);
-        return redirect()->back()->with('success', 'Программа была успешно удалена!');
+        if (Auth::user()->hasPermissionTo('program-revoke')) {
+            $userId = filter_var($request->user, FILTER_SANITIZE_NUMBER_INT);
+            $program = filter_var($request->program, FILTER_SANITIZE_SPECIAL_CHARS);
+            $user = User::findOrFail($userId)->programs()->detach($program);
+            return redirect()->back()->with('success', 'Программа была успешно удалена!');
+        } else {
+            return redirect()->back()->with('error', 'У Вас нет прав для выполнения этой операции');
+        }
     }
 }
